+++ build/deploy_for_web.bat
@@ -8,12 +8,14 @@ rem Kuin2.kn ->[Kuin1.exe]-> Kuin2.exe
 rem ---------------------------------------------------------------------------
 
 rem Kuin2.kn ->[Kuin2.exe]-> Kuin2.js
-copy /Y "..\src\compiler\*.kn" "..\src\compiler\res\"
-xcopy /s /e /q /i /y "..\src\compiler\cpp" "..\src\compiler\res\cpp"
-xcopy /s /e /q /i /y "..\src\compiler\web" "..\src\compiler\res\web"
-rd /s /q "%~dp0deploy_web"
+if exist "%~dp0deploy_web_res" rd /s /q "%~dp0deploy_web_res"
+mkdir "%~dp0deploy_web_res"
+copy /Y "..\src\sys\common.h" ".\deploy_web_res\"
+xcopy /s /e /q /i /y "..\src\sys\cpp" ".\deploy_web_res\cpp"
+xcopy /s /e /q /i /y "..\src\sys\web" ".\deploy_web_res\web"
+if exist "%~dp0deploy_web" rd /s /q "%~dp0deploy_web"
 mkdir "%~dp0deploy_web"
-.\output\kuin.exe -i "%~dp0../src/compiler/main.kn" -o "%~dp0deploy_web/kuin" -s "%~dp0../src/compiler/res/sys/" -e web -x static
+.\output\kuin.exe -i "%~dp0../src/compiler/main.kn" -o "%~dp0deploy_web/kuin" -s "%~dp0../src/sys/" -p "%~dp0./deploy_web_res/" -e web -x static -r
 
 pause
